refactor: extract application builder
